# shoes/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Sample CRUD api for shoes_development db",
        "title": "Sample CRUD api for shoes_development db",
        "contact": {
            "name": "Me",
            "url": "http://me.com/terms.html",
            "email": "me@me.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "xinqi.dev:443",
    "basePath": "/",
    "paths": {
...
        "/schemamigrations_": {
            "get": {
                "description": "GetAllSchemaMigrations_ is a handler to get a slice of record(s) from schema_migrations table in the shoes_development database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SchemaMigrations_"
                ],
                "summary": "Get list of SchemaMigrations_",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page requested (defaults to 0)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "number of records in a page  (defaults to 20)",
                        "name": "pagesize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "db sort order column",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.PagedResults"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "description": "add to add a single record to schema_migrations table in the shoes_development database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SchemaMigrations_"
                ],
                "summary": "Add an record to schema_migrations table",
                "parameters": [
                    {
                        "description": "Add SchemaMigrations_",
                        "name": "SchemaMigrations_",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SchemaMigrations_"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.SchemaMigrations_"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            }
        },
        "/schemamigrations_/{argVersion}": {
            "get": {
                "description": "GetSchemaMigrations_ is a function to get a single record from the schema_migrations table in the shoes_development database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SchemaMigrations_"
                ],
                "summary": "Get record from table SchemaMigrations_ by  argVersion",
                "operationId": "argVersion",
                "parameters": [
                    {
                        "type": "string",
                        "description": "version",
                        "name": "argVersion",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.SchemaMigrations_"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "404": {
                        "description": "ErrNotFound, db record for id not found - returns NotFound HTTP 404 not found error",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a single record from schema_migrations table in the shoes_development database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SchemaMigrations_"
                ],
                "summary": "Update an record in table schema_migrations",
                "parameters": [
                    {
                        "type": "string",
                        "description": "version",
                        "name": "argVersion",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update SchemaMigrations_ record",
                        "name": "SchemaMigrations_",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SchemaMigrations_"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.SchemaMigrations_"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a single record from schema_migrations table in the shoes_development database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SchemaMigrations_"
                ],
                "summary": "Delete a record from schema_migrations",
                "parameters": [
                    {
                        "type": "string",
                        "description": "version",
                        "name": "argVersion",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/model.SchemaMigrations_"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        },
        "api.PagedResults": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total_records": {
                    "type": "integer"
                }
            }
        },
...
        "model.SchemaMigrations_": {
            "type": "object",
            "properties": {
                "version": {
                    "description": "[ 0] version                                        varchar(255)         null: false  primary: true   isArray: false  auto: false  col: varchar         len: 255     default: []",
                    "type": "string"
                }
            }
        },
...
    }
}